refactor validation tools
